+++ libc/hexdump.c
@@ -5,7 +5,7 @@
 static char dump_byte_char(char c) { return isprint(c) ? c : '.'; }
 
 static void print_byte_char_line(const char *c, size_t remaining_len) {
-	for (int i = 0; i < remaining_len; i++) {
+	for (size_t i = 0; i < remaining_len; i++) {
 		printf("%c", dump_byte_char(c[i]));
 	}
 }
@@ -31,7 +31,7 @@ static void hexdump_line(
 
 void hexdump_addr(const void *data, size_t len, uintptr_t base_address) {
 	for (size_t i = 0; i < len; i += 16) {
-		hexdump_line(data + i, MIN(len - i, 16), base_address + i);
+		hexdump_line(data + i, MIN(len - i, 16ul), base_address + i);
 	}
 }
